Support buggy .zips containing backslashes in the filename
There are archives in the wild made with buggy compressors that use
backslash as path separator. The ZIP format explicitly denies the
use of "\" so we just replace it with slash "/".

Moreover it seems that folders are stored as "files" but with a final
"\" in the filename, in this case we force directory creation even
if the entry is stored as "file".