Allow overwriting symlinks
When extracting regular files, the file is opened for writing with
os.O_TRUNC, making it suitable to extract over an existing file. This
allows updating a local copy from an updated archive by extracting over
it. If a link is added to the archive though, the second time it's
extracted, an error occurs since os.Link and os.Symlink will not
overwrite. This removes any file currently in place before attempting to
write a link.